requirements.txt
PyPDF2==3.0.1
flask==3.0.0
openai>=1.86.0
python-dotenv==1.0.0
markdownify==0.11.6
pytesseract==0.3.10
Pillow==10.0.1
pdf2image==1.16.3
extract-msg==0.47.0
email-reply-parser==0.5.12
python-magic==0.4.27
python-magic-bin==0.4.14
